requirements.txt
Flask==2.3.3
openai==1.35.7
python-dotenv==1.0.0
pylint==2.17.4
flake8==6.0.0
radon==6.0.1
tenacity==8.2.3
plotly==5.15.0
pytest==7.4.0
httpx==0.27.0